The Basics
I've created the basic layout and accessibility of the website. Right
now, the only connection to the database that's made is a check to see
if it exists, and a creation of it if it does not.

Username and passwords accept anything. There is no specification yet
for the username choosing which password/create password field pops up,
so both come up presently, and accept anything (for the means of getting
access to the webpage).

The options>help choice on the top right and choosing the day of the
week both work as well. Adding a refresh capability is currently not
present; I will explore adding that in once I've got everything I've
already learned implemented.