Stopped using !important
Surprisingly short search showed me something that I hadn't discovered
before, and the post explained the weights of classes and such, and how
to be more specific than bootstrap. I don't know if it's bad coding or
not to add a class or id to the body tag, it did seem the most ideal for
overriding any part of the page body though.