The user can now log in securely
Two different encryptions are used, one client side and one server side.
The one client side is to make sure others can't find the password in
the client-server communication and type it in. The server side is for
the people that do pick up the server-client encryption so they can't
bypass the server and just access the database. as easily for someone's
stuff.
Seemed the best way to do it with the tools I had.

Logging in, creating user, and returning users are all set up now.