Splitting the Javascript for readability
Extra files were made to make the code more organized, rather than
having everything in one big central mess. For the scale of this
project, this should work; in future, however, there is a way to
compress all of the code into one file (possibly server operation), and
have that file be referenced by the webpage. Possibly will look into
that.