Logout & delete
Allows user to log out or delete their account. In addition, there is
now an error space for the task bar, as well as 'back' buttons in the
password fields to allow the user to change their email if they've
mistyped.

Code for user logout/delete to be split from main.js in next commit.